+++ WDE.SmartScriptEditor/Editor/ISmartScriptImporter.cs
@@ -8,7 +8,7 @@ namespace WDE.SmartScriptEditor.Editor
     public interface ISmartScriptImporter
     {
         Task Import(SmartScript script, bool doNotTouchIfPossible, IReadOnlyList<ISmartScriptLine> lines, IReadOnlyList<IConditionLine> conditions, IReadOnlyList<IConditionLine>? targetConditions);
-        Dictionary<int, List<SmartCondition>> ImportConditions(SmartScriptBase script, IReadOnlyList<IConditionLine> lines);
+        Dictionary<long, List<SmartCondition>> ImportConditions(SmartScriptBase script, IReadOnlyList<IConditionLine> lines);
         List<SmartCondition> ImportConditions(SmartScriptBase script, IReadOnlyList<ICondition> lines);
     }
 }
