+++ WoWDatabaseEditor.Common/WDE.Common/Debugging/Events/ConditionBreakpointHitArgs.cs
@@ -6,7 +6,7 @@ public class ConditionBreakpointHitArgs
 {
     public int SourceType { get; init; }
     public int SourceEntry { get; init; }
-    public int SourceGroup { get; init; }
+    public long SourceGroup { get; init; }
     public int SourceId { get; init; }
     public AbstractCondition Condition { get; init; } = new();
 }
